pyproject.toml
[tool.poetry]
name = "django_bike"
version = "1.0.0"
description = ""
authors = ["Martin Lehoux <martin@lehoux.net>"]

[tool.poetry.dependencies]
python = "^3.8"
celery = "^4.4"
requests = "^2.25"
redis = "^3.5"
django-dotenv = "^1.4"
plotly = "^4.14"
pandas = "^1.2"
psycopg2-binary = "^2.8"
gunicorn = "^20.0"
django-sendgrid-v5 = "^0.8"
django-cleanup = "^5.1"
channels = "^3.0"
channels_redis = "^3.2"
pytest = "^5.4"
pytest-django = "^3.10"
pytest-cov = "^2.11"
django_coverage_plugin = "^1.8"
pytest-asyncio = "^0.14"
django-redis = "^4.12"
Pillow = "^8.1"
PyYAML = "^5.4"
gpxpy = "^1.4"
"SRTM.py" = "^0.3.7"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
flake8 = "^3.9"
flower = "^0.9"
django-debug-toolbar = "^3.2"
isort = "^5.7"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.isort]
profile = "black"
